Add files via upload
Python note book files that demonstrate stepwise development of the Extended Variable Parallel Tempering (EVPT) method that I tested and used to perform a lot of CSP sampling in the 7th blind test I am currently still developing this method as a means of biassing a CSP search for elucidating crystal structure models from PXRD without the need for unit-cell indexing. The work remains unfunded and has been placed as a low priority.